Fix build warning in vld_main for truncation of size_t to int

+++ src/tests/vld_main/vld_main.cpp
@@ -11,7 +11,7 @@
 
 class MemoryLeak {
 public:
-    MemoryLeak(size_t n) { l = malloc(n); memset(l, 0x30 + (n / 10), n); } // 4,5
+    MemoryLeak(size_t n) { l = malloc(n); memset(l, 0x30 + ((int)n / 10), n); } // 4,5
     ~MemoryLeak() { free(l); }
 private:
     void* l;
